remove gitRoot option and throw error for relative projectRoot
