+++ packages/core/cli/src/normalizeOptions.ts
@@ -36,7 +36,6 @@ export interface Options {
   cache?: boolean;
   cacheDir?: string;
   projectRoot?: string;
-  gitRoot?: string;
   watchDir?: string;
   watchBackend?:
     | 'watchman'
@@ -180,7 +179,6 @@ export async function normalizeOptions(
     shouldDisableCache: command.cache === false,
     cacheDir: command.cacheDir,
     projectRoot: command.projectRoot,
-    gitRoot: command.gitRoot,
     watchDir: command.watchDir,
     watchBackend: command.watchBackend,
     watchIgnore: command.watchIgnore,
