+++ packages/core/core/src/resolveOptions.ts
@@ -102,12 +102,12 @@ export default async function resolveOptions(
   }
 
   let projectRoot;
-  if (initialOptions.projectRoot != null) {
-    // Use explicitly provided projectRoot, resolving relative paths to cwd
-    const providedProjectRoot = initialOptions.projectRoot;
-    projectRoot = path.isAbsolute(providedProjectRoot)
-      ? providedProjectRoot
-      : path.resolve(inputCwd, providedProjectRoot);
+  if (initialOptions.projectRoot) {
+    // Use explicitly provided projectRoot
+    if (!path.isAbsolute(initialOptions.projectRoot)) {
+      throw new Error('Specified project root must be an absolute path');
+    }
+    projectRoot = initialOptions.projectRoot;
   } else {
     // getRootDir treats the input as files, so getRootDir(["/home/user/myproject"]) returns "/home/user".
     // Instead we need to make the the entry refer to some file inside the specified folders if entries refers to the directory.
@@ -125,16 +125,7 @@ export default async function resolveOptions(
     projectRoot = path.dirname(projectRootFile);
   }
 
-  let gitRoot;
-  if (initialOptions.gitRoot != null) {
-    const providedGitRoot = initialOptions.gitRoot;
-    gitRoot = path.isAbsolute(providedGitRoot)
-      ? providedGitRoot
-      : path.resolve(inputCwd, providedGitRoot);
-  } else {
-    gitRoot = await findGitRepositoryRoot(inputFS, projectRoot);
-  }
-
+  const gitRoot = await findGitRepositoryRoot(inputFS, projectRoot);
   if (inputFS instanceof NodeVCSAwareFS) {
     inputFS.setGitRepoPath(gitRoot);
   }
